refactor: class moved to seperate file and renamed
