+++ shared/src/commonMain/kotlin/data/bletransfer/Holder.kt
@@ -1,7 +1,7 @@
 package data.bletransfer
 
 import androidx.compose.runtime.Composable
-import data.bletransfer.verifier.ReceivedDocument
+import data.bletransfer.holder.RequestedDocument
 
 expect fun getHolder(): Holder
 
@@ -9,7 +9,7 @@ interface Holder {
     @Composable
     fun getRequirements(check: (Boolean) -> Unit)
 
-    fun hold(updateQrCode: (String) -> Unit, updateRequestedAttributes: (List<ReceivedDocument>) -> Unit)
+    fun hold(updateQrCode: (String) -> Unit, updateRequestedAttributes: (List<RequestedDocument>) -> Unit)
 
     fun disconnect()
 }
